initial update
The initial update contains two raw text files, that can be edited by all ASCII Text-Editors like notepad++ or nano.
The contents are used, to render a PDF file in a Adobe PDF Viewer (this can be done by a HTML-Browser lile Edge or FireFox, ...).
I have use simple constructs, that can be found on several locations in the www.
Have fun, and report me, what you have done with, based on this github stuff.
Thanks for reading, paule32